+++ sideline/components/playbook/PlaySheetActiveBadge.tsx
@@ -1,6 +1,0 @@
-import { PLAY_SHEET_ACTIVE_BADGE } from "@/lib/coachCopy";
-import { playSheetActiveBadgeClass } from "@/lib/constants/designTokens";
-
-export function PlaySheetActiveBadge() {
-  return <span className={playSheetActiveBadgeClass}>{PLAY_SHEET_ACTIVE_BADGE}</span>;
-}